backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
sentence-transformers>=2.5.0
faiss-cpu>=1.7.0
numpy>=1.24.0
openai>=1.12.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
flake8>=7.0.0
